fix: Draw view-hierarchy's Surface on top of SurfaceView's
Issue related to adding a "presentation box" to the top-left of the screen (sitting above BOTH the AnimatedLogoView and the camera's viewport). Unable to "sandwich" the camera's viewport (a SurfaceView's Surface) between the AnimatedLogoView and the "presentation box" (both of which are Views from the regular view-hierarchy [which gets composited into one Surface, separate from the SurfaceView's Surface).