feat: Replace AnimatedLogoView with TextViews-with-marquee-functionality
Instead of a custom view with its properties animated by an ObjectAnimator, there is a RecyclerView with a vertical list of TextViews configured to use its built-in marquee functionality (the adapter for the RV has 2 ViewHolder types [one with a LTR moving TextView/marquee and the other with a RTL moving TextView/marquee]).

Cleaned up project (remove AnimatedLogoView and the test Fragment used in the previous commit.