tinker: Adjust logic for more responsive control
Instead of DIRECTLY using the device's CURRENT accelerometer values to determine the player's velocity (which determines the change of distance), use the DIFFERENCE between the CURRENT and PREVIOUS values.

The former approach requires the user to tilt the device pass the midpoint/zero-value of the axis of interest in order to change the player's direction.

This approach changes the player's direction using the change between the present and previous acceleration.