tinker: Add MilkSteamingPitcher and MilkTouchListener
Start drag and drop event when views-representing-boxes-of-milk from RefrigeratorFragment are touched.

Add MilkSteamingPitcher to MaestranaFragment's framelayoutRight. The current touch listener is just a test that iterates through possible types of milk that the pitcher can contain (or [null] if pitcher is empty).

NEXT: implement drag listener to handle drop-target for dragging milk from RefrigeratorFragment.