tinker: Change text color on espresso shot collision
Espresso shot collision (with CupImageView and ShotGlass) does more than increment the numberOfShots field, it now tracks the type of espresso (which specifies which text color to use in onDraw()).

NEXT: pour/transfer content from ShotGlass to CupImageView (similar to SteamingPitcher's content to CupImageView) (the current goal is to craft a latte).