tinker: Handle new data (R, L, 1/2, U) resulted from EspressoShotControlDialogFragment
The data configures the EspressoShot (amountOfWater and amountOfBean). This data is transferred from the EspressoShot to either a ShotGlass or CupImageView if there is a View-to-View collision. ShotGlass and CupImageView can also transfer this data between each other.

Currently, have not used this new data for CupImageView's drag handler of DrinkLabel (checking for winner).