refactor: Re-work shaking logic for IceShaker
Previously, a successful up-and-down repetition was based on distance from the y-value at the start of the drag.

Now, an up-and-down repetition is base on distance from the most recent yPeak or yTrough (a shake no longer revolve around the starting y-value, a shake can now be successful irregardless of the starting y-value).