fix: Use correct coordinate values to separate horizontal/vertical movements
This version uses the correct values when checking if a Tile is solid or not (depending on the direction of movement). Since this version uses the correct values, it is the first version where horizontal and vertical movements are independent. Still using a reduced-by-1 (depending on direction of movement) collision bound values.