tinker: Make horizontal/vertical movements independent for entity collision
The previously declared "first version where horizontal and vertical movements are independent" was true only for tile collision detection.

This version is the "first version where horizontal and vertical movements are independent" for BOTH entity AND tile collision detection.

The player no longer gets "stuck" if one of its side is adjacent to another entity.