tinker: Change scene on transfer point collision
Transfer point is currently 2-tile above and 1-tile right of the player's spawn position in WorldScene. LabScene's transfer point is 1-tile below the player's spawn position.

Returning from LabScene (back to WorldScene) did not work as expected. After a lot of narrowing down the bug, it appears that WorldScene.enter() isn't having an effect on the player though it is being called when changing scenes. To work-around enter() not having an effect, I do scene "entering" tasks inside a conditional in update().