feat: Implement transfer point cool-down system
Entering a scene will restart that scene's cool-down timer, which limits calling changeScene() to only when the "cool-down threshold" is reached.

Previously, scene changes had the potential condition of changing back and forth.

This approach uses time. Another approach could've been to specify a tile-adjacent-to-the-transfer-point based on the player's direction during collision.