fix: Adjust tile in front of player's house (SceneFarm) to NOT be GrowableTile
This will prevent an entity-collision-lock. Previously, SceneFarm.startNewDay() could've placed Plant there (the tile in front of the player's house is its transfer point spawn position) when it was a GrowableTile.

Move GrowableTile's code related to seed germination into its own method. Use this method during SceneFarm.init() to set-up a nearly harvestable Plant on the tile that's 2-steps away from the front of player's house.