tinker: Add 6th button (harvest) to RobotDialogFragment
TileSelectorDialogFragment is now compatible with selecting tiles that have Plant in a harvestable state (when the mode is set to HARVEST).

Robot will pick-up the Carryable (Plant in harvestable state), change its direction to UP, and place-down the Carryable onto the tile above it.

Change robot's movement animator's duration to be RUNNING_MOVEMENT_DURATION and reduced robot's COUNTER_COMMANDS_TARGET from 50 to 25 (now it'll take half the time to perform a task from its command queue).