tinker: Use changing-bubble-colors as visual indicator of impending escape
Monsta that have been bubbled will be green during the first half of the escape timer incrementing towards its target.

During the second half of reaching the timer's target, the bubbled-Monsta will randomly change between blue, orange, and red.

When the timer reaches its target, the Monsta escapes the bubble and patrols to the right.