refactor: Compose Bubblun w BubblunStateManager
Bubblun's animation changes from IDLE to ATTACK when a-button is pressed (and Bubblun is the entity in front of the player). It changes back to IDLE after the ATTACK animation finishes.

The end result is Bubblun opens his mouth when put into ATTACK state, and closes his mouth when ATTACK state ends (going back to IDLE state).

Attacking and moving/jumping should be able to happen simultaneously, so BubblunStateManager should account for both in each call to update().