feat: Show quest log when quest icon is clicked
Move logic of StatsDisplayerFragment's icon clicks to Game class.

Clicking the quest icon shows a FCVDialogFragment (its width and height can be defined as a percentage of the screen's width and height) that contains a fragment. In this case, the fragment is a RecyclerViewFragment (a fragment with a TextView [title] above a RecyclerView [list]).

The RecyclerView's items have selectable background that gives a visual cue when clicked. The RecyclerView's items also have click listener (currently place-holders).