feat: Implement "side-conversation" for NonPlayableCharacter who are a part of the player's party
In WorldScene, player's collision with "Youngster" (the 3rd NPC from bottom of nugget bridge) causes him to join the player's party.

"Youngster" follows the party leader's path, and triggers a line of conversation on every 3rd step (then it disappears on the next 3rd step, changes from left-to-right/right-to-left).

Besides TelevisionFragment....... this will be the MAIN ROUTE of DISSEMINATING KNOWLEDGE/FRONT-LOADING.