feat: Add second quest (new Item added [RobotReprogrammer4000])
Player no longer able to open RobotDialogFragment by simply walking up to the robot and clicking a-button.

RobotDialogFragment can only be launched using RobotReprogrammer4000 (a new Item that implements EntityCommandOwner).

The RobotReprogrammer4000 is obtained after finishing the first quest, talking to the seed shop owner, exiting the seed shop, and waiting for the AI's dialogue to finish (it is a starting item from AIQuest00).