tinker: Implement single-click, long-click, and double-click for ProjectViewportFragment
Single-click on an item from ProjectViewportFragment changes the fragment inside StructureViewportFragment.

Long-click on an item from ProjectViewportFragment shows an EditText to rename the class (which changes the name shown in MainViewportFragment's tab and associated fragment).

Double-click on an item from ProjectViewportFragment adds that class to MainViewportFragment's list of currently opened classes.