tinker: Display Main class in ClassEditorFragment to be superficially accurate
It has proper indentations, return type, method name, argument list (round-braces), and empty body (curly-braces).

FIX: when opening a class via double-click of an item from ProjectViewportFragment, there is now a check for whether the class is already opened (if so, merely show its fragment in MainViewportFragment, otherwise add it to the list AND then show it).