tinker: Add third quest (RobotDialogQuest00)
The second quest is to till 4,200 GrowableTile (by either brute-force with the Shovel item or reprogramming the Robot to till).

This third quest is to open the Robot's IDEDialogFragment.

NEXT: add RobotDialogQuest01 to double click Robot class from ProjectViewportFragment (which opens it in a ClassDisplayerFragment [from MainViewportFragment's VP2]). Afterward, have another quest to go back to Main class and learn looping constructs or something.