fix: Continue conversation on new "page" if too long for one
No pressing of button to continue on to next page, merely starting a new "page" when the animation finishes and there are more pages left to show.

Conversations in WorldScene/LabScene and SceneFarm/SeedShopDialogFragment no longer runs off the screen vertically.

Increased the delay of text reveal (slowed the text appearance rate).