tinker: Replace NextWeekTonight's SurfaceView with VideoView
Replaced the use of the device's camera on a SurfaceView for a resource video played on a VideoView.

The NextWeekTonightFragment with the device's camera is preserved in a new class named NextWeekTonightUsingCameraFragment.

The new version of NextWeekTonightFragment (the one using a VideoView) is now named NextWeekTonightEpisodesGeneratorFragment.

This name is more suitable because I plan to use my phone's "screen recorder" feature to record this fragment in action while using OnClickListener to change the top-left panel (the "presentation box"). This is an alternative to complicated syncronizing/timing logic.