tinker: Change over-watering to diseased
Add Scissors item to remove any Plant (diseased or not).

Player is suppose to use PlantInspectioner200 to check if the plant is diseased or not, then remove it using the scissors item (no visual hints [unlike watering]).

This is intended to make the repeated task tedious, while the automation using a reprogrammed robot makes the repeated task quick.