tinker: Change highlighting/unhighlighting from single tile to list of tiles
Highlighting/unhighlighting functionality was thrown together as fast as possible just to have a visual to indicate which step I'm at in the DailyLoop.

Must re-do initial sizing and then re-do approach of drawing tile highlighting (previously had a small map where player moving didn't pan the game camera... now it's a larger map and this bug is apparent).