feat: Handle list of images for NextWeekTonightEpisodesGeneratorFragment
Previously, could only have images and videos inside a presentation box.

Now, if the resource is a list of images (instead of a video or a single image), it'll hide the presentation box and add the list of images (one per click) stacked on top of each other (but staggered so audience can still infer order/sequence) until it reaches the end of the list (then it'll show the presentation box).