feat: Create locale-specific resource directory (values-es)
Organize "strings.xml" and externalize user-facing text (so far just went through the starting part of the game).

Now, if phone's language-setting is set to spanish, the game uses the "strings.xml" from the (values-es) directory instead of the default (values) directory (i.e. game is starting to be presented in spanish).

FIX: limit clicking of StatsDisplayerFragment's icons to only when in GameState (instead of NewOrContinueState or IntroState).