fix: Handle robot's commands/tileWorkRequests differently
Previously, robot's position was calculated before the visual movement of the interpolator finished. This caused the robot to think it reached the end of it's walking one tile right (or left, up, down) too early.

Now, handling the next command/tileWorkRequest when the ObjectAnimator's onAnimationEnd() is called (rather than in update()).

Now, it's working properly, but the code is really messy and needs to be cleaned.